+++ CHANGELOG.md
@@ -185,6 +185,11 @@
 ### Features
 - Add possibility to delete custom fields and their options
 
+## v4.15.4 (2026-03-06)
+- Fix statement vote on mysql8+, immediately show vote
+- Rate limit new statements only for anonymous users
+- Set TUS resumable upload cache TTL to a week by default
+
 ## v4.15.3 (2025-12-02)
 
 ## v4.15.2 (2025-10-24)